Jingyi's work
I found there still are many missing data in the dataset. Because the number of vaccination must be an increasing trend, we cannot use mean or 0 to fill the empty. I learned and used interpolate method to fill them. Additionally, I graphed a line chart for people in all age ranges in six cities. By using the cleaned data and the plot method I used, our group can graph the trends in vaccination numbers for any kind of group range and any city.