Jingyi's work
I found there still are many of missing data in dataset. Because the number of vaccination must be in an increasing trend, we cannot use mean or 0 to fill the empty. I learned and used interpolate method to fill them. Additionally, I graphed a line chart for people in all age ranges in six cities. Our group can use this cleaned dataset and the plot method I used to graph the trend of vaccination numbers for any age range and all six cities.